dispatchorder datatable improv.

+++ app/Http/Livewire/DispatchOrders/Form.php
@@ -52,6 +52,7 @@ class Form extends Component
     public function mount($dispatchOrder = null)
     {
         if($dispatchOrder) {
+            if($dispatchOrder->isFinalized()) abort(404);
             $this->setEditMode($dispatchOrder);
         } else {
             $this->do_planned_datetime = Carbon::today();
@@ -90,6 +91,8 @@ class Form extends Component
         $this->validate($this->spRules);
         
         if($this->editMode === true) {
+            if($this->dispatchOrder->isFinalized()) abort(404);
+            
             $this->dispatchOrder->dispatchProducts()->delete();
             
             $this->dispatchOrder->update($validatedDoData);
